requirements.txt
azure-ai-textanalytics==5.1.0
azure-cognitiveservices-speech==1.20.0
azure-common==1.1.28
azure-core==1.23.1
azure-identity==1.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
docopt==0.6.2
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
pipreqs==0.4.11
portalocker==2.4.0
pycparser==2.21
PyJWT==2.3.0
pywebio==1.6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tornado==6.1
typing_extensions==4.1.1
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
yarg==0.1.9